+++ cmd/ocm-support/delete/registry_credentials/cmd.go
