+++ cmd/ocm-support/get/registry_credentials/cmd.go
